feat: add useFacetsConfig hook for improved facets configuration handling
