+++ keep-ui/features/filter/facet-panel-server-side.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useState } from "react";
 import {
   FacetDto,
   FacetOptionsQueries,
