+++ keep-ui/features/filter/store/index.ts
@@ -1,2 +1,3 @@
 export * from "./use-store";
 export { useFacetsLoadingStateHandler } from "./use-facets-loading-state-handler";
+export { useFacetsConfig } from "./use-facets-config";
